requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
python-Wappalyzer>=0.3.1
colorama>=0.4.6
tqdm>=4.66.1
aiohttp>=3.9.1
asyncio>=3.4.3
prettytable>=3.9.0
pyyaml>=6.0.1
cryptography>=42.0.5
pyOpenSSL>=24.0.0 